pyproject.toml
[project]
name = "ai-jobs-project"
version = "0.1.0"
description = ""
authors = [
    {name = "swhh",email = "seamus.holland1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth (>=2.40.3,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.172.0,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "google-genai (>=1.21.1,<2.0.0)",
    "validators (>=0.35.0,<0.36.0)"
]

[tool.poetry]
packages = [{include = "ai_jobs_project", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
